+++ src/services/CoreService.php
@@ -1,12 +1,11 @@
 <?php
 /**
- * Webtexttool plugin for Craft CMS
+ * Textmetrics plugin for Craft CMS
  *
  * @author    Israpil Nalgiev
- * @link      https://www.webtexttool.com/
- * @copyright Copyright (c) 2018 Israpil Nalgiev
+ * @link      https://www.textmetrics.com/
+ * @copyright Copyright (c) 2019 Textmetrics
  */
-
 namespace inalgiev\webtexttool\services;
 
 use inalgiev\webtexttool\records\CoreRecord;